Updated Folder structure to better conform with standards
