+++ Pensopay/Models/Settlements/Settlement.cs
@@ -1,11 +1,6 @@
 ﻿using Pensopay.Models.Util.Settlements;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Pensopay.Models
+namespace Pensopay.Models.Settlements
 {
     public class Settlement
     {
